chore(firmware): extract FirmwareUpdateCoordinator from DaqifiViewModel (#592)
Move the PIC32 + WiFi firmware update orchestration and the firmware
version-check / outdated-device notification logic out of the 2,254-line
DaqifiViewModel god class into a dedicated, unit-testable
FirmwareUpdateCoordinator (under Device/Firmware/, per CLAUDE.md).

This is the first behavior-preserving extraction called for in #592
(highest-value item: FirmwareUpdateCoordinator). DaqifiViewModel drops
~530 lines (2,254 -> 1,724) and keeps only the bound progress/status
properties and thin command bindings, so the firmware XAML is unchanged.

How it is wired:
- New IFirmwareUpdateHost seam exposes the view-model state the coordinator
  reads (SelectedDevice, FirmwareFilePath, connected devices) and writes
  (progress/status/flags, notifications). DaqifiViewModel implements it; its
  existing [ObservableProperty] members stay the binding source, so no
  XAML or DataContext changes are needed.
- The coordinator constructor-injects the firmware services (resolving the
  same production defaults the view model used to build) and never reaches
  into ConnectionManager.Instance directly.
- GetFirmwareupdatationList typo fixed -> RefreshFirmwareUpdatesAsync.

Tests: the existing firmware test asset
(DaqifiViewModelFirmwareUpdateTests.cs) is retargeted to drive the
coordinator directly through a lightweight fake host instead of a full
view model -- same behavior contract, now WPF-free. Added unit coverage
for the newly-testable version-check / notification path.

Verified: full solution builds; 372 unit tests pass
(TestCategory!=Ui&FullyQualifiedName!~WindowsFirewallWrapperTests); app
launches without startup regressions.

Follow-up extractions from #592 (logging-session list, disk-space monitor,
confirm overlay; DatabaseLogger split) remain separate PRs.
