+++ Daqifi.Desktop/ViewModels/DaqifiViewModel.cs
@@ -15,22 +15,17 @@ using MahApps.Metro.Controls.Dialogs;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions;
 using Daqifi.Core.Firmware;
-using Daqifi.Core.Communication.Transport;
 using Daqifi.Desktop.Device.Firmware;
 using Microsoft.Data.Sqlite;
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 using System.ComponentModel;
 using System.IO;
-using System.Net.Http;
 using System.Windows;
 using System.Windows.Input;
 using System.Windows.Threading;
 using CommunityToolkit.Mvvm.ComponentModel;
-using Daqifi.Desktop.Device.SerialDevice;
-using System.IO.Ports;
 using Application = System.Windows.Application;
 using File = System.IO.File;
 using CommunityToolkit.Mvvm.Input;
@@ -38,7 +33,7 @@ using Daqifi.Core.Device.SdCard;
 
 namespace Daqifi.Desktop.ViewModels;
 
-public partial class DaqifiViewModel : ObservableObject
+public partial class DaqifiViewModel : ObservableObject, IFirmwareUpdateHost
 {
     private readonly AppLogger _appLogger = AppLogger.Instance;
 
@@ -155,16 +150,12 @@ public partial class DaqifiViewModel : ObservableObject
     private string _firmwareUpdateStatusText = string.Empty;
     [ObservableProperty]
     private bool _selectedDeviceSupportsFirmwareUpdate;
-    private readonly IFirmwareUpdateService _firmwareUpdateService;
-    private readonly IFirmwareDownloadService _firmwareDownloadService;
     private readonly IDbContextFactory<LoggingContext>? _loggingContextFactory;
-    private readonly Func<string, string, IFirmwareUpdateService> _wifiFirmwareUpdateServiceFactory;
-    private CancellationTokenSource? _firmwareUploadCts;
+    private readonly FirmwareUpdateCoordinator _firmwareCoordinator;
     private ConnectionDialogViewModel _connectionDialogViewModel;
     private string _selectedLoggingMode = "Stream to App";
     private bool _isLogToDeviceMode;
     private SdCardLogFormat _selectedSdCardLogFormat = SdCardLogFormat.Protobuf;
-    private IStreamingDevice? _deviceBeingUpdated;
     private IDiskSpaceMonitor? _diskSpaceMonitor;
     private ObservableCollection<LoggingSession>? _observedLoggingSessions;
     private CancellationTokenSource? _networkSettingsAppliedCts;
@@ -538,11 +529,19 @@ public partial class DaqifiViewModel : ObservableObject
         IDbContextFactory<LoggingContext>? loggingContextFactory = null)
     {
         _dialogService = dialogService ?? throw new ArgumentNullException(nameof(dialogService));
-        _firmwareDownloadService = firmwareDownloadService ?? CreateDefaultFirmwareDownloadService();
-        var resolvedLogger = firmwareLogger ?? NullLogger<FirmwareUpdateService>.Instance;
-        _firmwareUpdateService = firmwareUpdateService ?? CreateDefaultFirmwareUpdateService(_firmwareDownloadService, resolvedLogger);
         _loggingContextFactory = loggingContextFactory;
-        _wifiFirmwareUpdateServiceFactory = wifiFirmwareUpdateServiceFactory ?? CreateWifiFirmwareUpdateService;
+
+        // Firmware update orchestration and version-checking live in the coordinator (issue #592).
+        // It resolves the production defaults itself when the injected services are null, mirroring
+        // the previous in-view-model construction; this view model keeps only the bound properties
+        // and thin command bindings, and feeds the coordinator through the IFirmwareUpdateHost seam.
+        _firmwareCoordinator = new FirmwareUpdateCoordinator(
+            this,
+            _dialogService,
+            firmwareUpdateService,
+            firmwareDownloadService,
+            firmwareLogger,
+            wifiFirmwareUpdateServiceFactory);
 
         ConfirmAffirmativeCommand = new RelayCommand(() => CompleteConfirm(true));
         ConfirmNegativeCommand = new RelayCommand(() => CompleteConfirm(false));
@@ -674,481 +673,14 @@ public partial class DaqifiViewModel : ObservableObject
     #region Updload firmware and update processes
 
     [RelayCommand]
-    private async Task UploadFirmware()
-    {
-        if (IsFirmwareUploading)
-        {
-            return;
-        }
-
-        if (SelectedDevice?.ConnectionType != Device.ConnectionType.Usb)
-        {
-            return;
-        }
-
-        if (SelectedDevice is not SerialStreamingDevice serialStreamingDevice)
-        {
-            return;
-        }
-
-        SelectedDeviceSupportsFirmwareUpdate = true;
-        HasErrorOccured = false;
-        IsUploadComplete = false;
-        UploadFirmwareProgress = 0;
-        UploadWiFiProgress = 0;
-        FirmwareUpdateStatusText = "Preparing firmware update...";
-
-        _deviceBeingUpdated = SelectedDevice;
-        ConnectionManager.Instance.DeviceBeingUpdated = _deviceBeingUpdated;
-
-        var isManualUpload = !string.IsNullOrWhiteSpace(FirmwareFilePath);
-
-        _firmwareUploadCts?.Dispose();
-        _firmwareUploadCts = new CancellationTokenSource();
-        IsFirmwareUploading = true;
-        _appLogger.AddBreadcrumb("firmware", $"Firmware update started for {serialStreamingDevice.Name}");
-
-        try
-        {
-            var coreDevice = serialStreamingDevice.ConnectedCoreStreamingDevice;
-
-            if (!coreDevice.IsConnected)
-            {
-                _appLogger.Error($"Device {serialStreamingDevice.Name} is not connected. Cannot update firmware on a disconnected device.");
-                NotificationList.Add(new Notifications
-                {
-                    Message = $"Please connect device {serialStreamingDevice.Name} before attempting firmware update.",
-                    DeviceSerialNo = serialStreamingDevice.DeviceSerialNo
-                });
-
-                return;
-            }
-
-            // For an auto-update, download into a LOCAL variable rather than the bound
-            // FirmwareFilePath property. Writing the downloaded path back into the property
-            // would make the NEXT in-session update look like a manual upload (isManualUpload
-            // is derived from FirmwareFilePath being non-empty), silently skipping the
-            // WiFi-module step until the app is restarted. See issue #599.
-            string effectiveFirmwarePath;
-            if (!isManualUpload)
-            {
-                FirmwareUpdateStatusText = "Downloading latest firmware package...";
-                effectiveFirmwarePath = await _firmwareDownloadService.DownloadLatestFirmwareAsync(
-                    GetFirmwareDownloadDirectory(),
-                    includePreRelease: true,
-                    cancellationToken: _firmwareUploadCts.Token);
-            }
-            else
-            {
-                effectiveFirmwarePath = FirmwareFilePath;
-            }
-
-            if (string.IsNullOrWhiteSpace(effectiveFirmwarePath) || !File.Exists(effectiveFirmwarePath))
-            {
-                throw new FileNotFoundException("Firmware file path is invalid or does not exist.", effectiveFirmwarePath);
-            }
-
-            var pic32Progress = new Progress<FirmwareUpdateProgress>(report =>
-            {
-                UploadFirmwareProgress = Math.Clamp((int)Math.Round(report.PercentComplete), 0, 100);
-                if (!string.IsNullOrWhiteSpace(report.CurrentOperation))
-                {
-                    FirmwareUpdateStatusText = report.CurrentOperation;
-                }
-            });
-
-            await _firmwareUpdateService.UpdateFirmwareAsync(
-                coreDevice,
-                effectiveFirmwarePath,
-                pic32Progress,
-                _firmwareUploadCts.Token);
-
-            if (!isManualUpload)
-            {
-                await UpdateWifiModuleAsync(coreDevice, serialStreamingDevice, _firmwareUploadCts.Token);
-            }
-
-            IsUploadComplete = true;
-            _appLogger.AddBreadcrumb("firmware", "Firmware update completed");
-            ShowUploadSuccessMessage();
-        }
-        catch (OperationCanceledException)
-        {
-            FirmwareUpdateStatusText = "Firmware update canceled.";
-            _appLogger.Warning("Firmware update canceled by user.");
-            _appLogger.AddBreadcrumb("firmware", "Firmware update cancelled", Common.Loggers.BreadcrumbLevel.Warning);
-        }
-        catch (FirmwareUpdateException ex)
-        {
-            _appLogger.AddBreadcrumb("firmware", $"Firmware update failed: {ex.FailedState}", Common.Loggers.BreadcrumbLevel.Error);
-            HandleFirmwareUpdateException(ex);
-        }
-        catch (Exception ex)
-        {
-            HasErrorOccured = true;
-            _appLogger.Error(ex, "Problem Uploading Firmware");
-            _appLogger.AddBreadcrumb("firmware", "Firmware update failed", Common.Loggers.BreadcrumbLevel.Error);
-            ShowFirmwareErrorDialog("Firmware update failed. Please try again.");
-        }
-        finally
-        {
-            IsFirmwareUploading = false;
-            _firmwareUploadCts?.Dispose();
-            _firmwareUploadCts = null;
-            _deviceBeingUpdated = null;
-            ConnectionManager.Instance.DeviceBeingUpdated = null;
-
-            // Consume any manual .hex selection so the auto/manual decision is a per-run input,
-            // not sticky session state. A manual upload is intentionally PIC32-only (no WiFi),
-            // and isManualUpload is derived from FirmwareFilePath being non-empty. Without this
-            // reset a prior manual selection would trap every later run in manual mode and
-            // silently skip the WiFi-module flash until the app is restarted — the symmetric
-            // case of issue #599. The next run defaults to a full auto-update unless the user
-            // explicitly re-selects a file. (The auto path never writes this property.)
-            FirmwareFilePath = string.Empty;
-        }
-    }
+    private Task UploadFirmware() => _firmwareCoordinator.UploadFirmwareAsync();
 
     [RelayCommand(CanExecute = nameof(CanCancelFirmwareUpload))]
-    private void CancelFirmwareUpload()
-    {
-        if (!IsFirmwareUploading)
-        {
-            return;
-        }
-
-        FirmwareUpdateStatusText = "Canceling firmware update...";
-        _firmwareUploadCts?.Cancel();
-    }
+    private void CancelFirmwareUpload() => _firmwareCoordinator.CancelUpload();
 
     private bool CanCancelFirmwareUpload()
     {
         return IsFirmwareUploading;
-    }
-
-    private async Task UpdateWifiModuleAsync(
-        Daqifi.Core.Device.IStreamingDevice coreDevice,
-        SerialStreamingDevice serialStreamingDevice,
-        CancellationToken cancellationToken)
-    {
-        // Let Core probe the WiFi version and look up the latest release in one call so the desktop
-        // can skip unnecessary downloads and surface the current/update versions in the UI. Core owns
-        // the startup retry policy for the chip-info probe (the chip may still be booting right after
-        // a PIC32 update; see FirmwareUpdateServiceOptions.LanChipInfoMaxAttempts/LanChipInfoRetryDelay).
-        // Because the desktop owns this check, the Core flash call below passes skipVersionCheck: true
-        // so the device isn't queried a second time.
-        const string versionUnavailableStatus = "WiFi firmware version unavailable; continuing with update.";
-
-        FirmwareUpdateStatusText = "Checking WiFi firmware version...";
-        _appLogger.Information("Checking WiFi firmware version before deciding whether to flash the WiFi module.");
-
-        WifiFirmwareStatus? wifiStatus = null;
-        try
-        {
-            wifiStatus = await _firmwareUpdateService.CheckWifiFirmwareStatusAsync(coreDevice, cancellationToken);
-        }
-        catch (OperationCanceledException)
-        {
-            throw;
-        }
-        catch (Exception ex)
-        {
-            // The status check is a UX optimization and expected failures already surface as Reason
-            // values, so this guards only unexpected faults (e.g. a broken service instance). Never
-            // let those abort the flash below, which runs on its own freshly created service.
-            _appLogger.Warning($"WiFi firmware status check failed ({ex.Message}); continuing with WiFi update.");
-            FirmwareUpdateStatusText = versionUnavailableStatus;
-        }
-
-        if (wifiStatus != null)
-        {
-            if (wifiStatus.CurrentChipInfo != null)
-            {
-                _appLogger.Information(
-                    "WiFi chip info query succeeded. " +
-                    $"Device WiFi firmware version: {wifiStatus.CurrentChipInfo.FwVersion}.");
-            }
-
-            switch (wifiStatus.Reason)
-            {
-                case WifiFirmwareStatusReason.UpToDate:
-                {
-                    var deviceVersion = wifiStatus.CurrentChipInfo!.FwVersion;
-                    var latestVersion = NormalizeWifiFirmwareVersion(wifiStatus.LatestRelease!.TagName);
-                    FirmwareUpdateStatusText = $"WiFi firmware already up to date ({deviceVersion}).";
-                    _appLogger.Information(
-                        $"WiFi firmware is already up to date (device: {deviceVersion}, latest: {latestVersion}); " +
-                        "skipping WiFi flash.");
-                    UploadWiFiProgress = 100;
-                    return;
-                }
-
-                case WifiFirmwareStatusReason.UpdateAvailable:
-                {
-                    var deviceVersion = wifiStatus.CurrentChipInfo!.FwVersion;
-                    var latestVersion = NormalizeWifiFirmwareVersion(wifiStatus.LatestRelease!.TagName);
-                    FirmwareUpdateStatusText =
-                        $"WiFi update available ({deviceVersion} → {latestVersion}). Downloading...";
-                    _appLogger.Information(
-                        $"WiFi firmware update required (device: {deviceVersion}, latest: {latestVersion}); " +
-                        "proceeding with WiFi flash.");
-                    break;
-                }
-
-                case WifiFirmwareStatusReason.ChipInfoUnavailable:
-                    _appLogger.Warning(
-                        "WiFi chip info unavailable after startup retries; continuing with WiFi update.");
-                    FirmwareUpdateStatusText = versionUnavailableStatus;
-                    break;
-
-                case WifiFirmwareStatusReason.DeviceDoesNotSupportLanQuery:
-                    _appLogger.Warning(
-                        "Device does not support WiFi chip info queries; continuing with WiFi update.");
-                    FirmwareUpdateStatusText = versionUnavailableStatus;
-                    break;
-
-                case WifiFirmwareStatusReason.LatestReleaseUnavailable:
-                    _appLogger.Warning(
-                        "Latest WiFi firmware release metadata was unavailable; continuing with WiFi update.");
-                    break;
-
-                default:
-                    // VersionUnparseable or future reasons — the comparison was inconclusive,
-                    // so conservatively continue with the flash.
-                    _appLogger.Warning(
-                        $"WiFi firmware version check was inconclusive ({wifiStatus.Reason}); " +
-                        "continuing with WiFi update.");
-                    FirmwareUpdateStatusText = versionUnavailableStatus;
-                    break;
-            }
-        }
-
-        FirmwareUpdateStatusText = "Downloading WiFi firmware package...";
-        var wifiDownloadProgress = new Progress<int>(percent =>
-        {
-            // Map download progress into the initial segment of the WiFi bar.
-            UploadWiFiProgress = Math.Clamp((int)Math.Round(percent * 0.2), 0, 20);
-        });
-
-        var wifiPackage = await _firmwareDownloadService.DownloadWifiFirmwareAsync(
-            GetWifiDownloadDirectory(),
-            wifiDownloadProgress,
-            cancellationToken);
-
-        if (wifiPackage == null)
-        {
-            throw new InvalidOperationException("No WiFi firmware package was found for update.");
-        }
-
-        var wifiVersion = NormalizeWifiFirmwareVersion(wifiPackage.Value.Version);
-        FirmwareUpdateStatusText = $"Updating WiFi module ({wifiVersion})...";
-        var wifiUpdateProgress = new Progress<FirmwareUpdateProgress>(report =>
-        {
-            UploadWiFiProgress = Math.Clamp((int)Math.Round(report.PercentComplete), 0, 100);
-            if (!string.IsNullOrWhiteSpace(report.CurrentOperation))
-            {
-                FirmwareUpdateStatusText = report.CurrentOperation;
-            }
-        });
-
-        // Preserve the legacy serial prep/reset sequence now that the firmware flow uses the
-        // underlying Core device directly instead of routing through a desktop-shaped adapter.
-        var lanUpdateModeEnabled = false;
-        try
-        {
-            lanUpdateModeEnabled = serialStreamingDevice.EnableLanUpdateMode();
-
-            var wifiUpdateService = _wifiFirmwareUpdateServiceFactory(wifiVersion, serialStreamingDevice.PortName);
-            try
-            {
-                await wifiUpdateService.UpdateWifiModuleAsync(
-                    coreDevice,
-                    wifiPackage.Value.ExtractedPath,
-                    wifiUpdateProgress,
-                    cancellationToken,
-                    skipVersionCheck: true);
-            }
-            finally
-            {
-                (wifiUpdateService as IDisposable)?.Dispose();
-            }
-        }
-        finally
-        {
-            if (lanUpdateModeEnabled)
-            {
-                serialStreamingDevice.ResetLanAfterUpdate();
-            }
-        }
-    }
-
-    private FirmwareUpdateService CreateWifiFirmwareUpdateService(string wifiVersion, string portName)
-    {
-        var firmwareLogger = App.ServiceProvider?.GetService<ILogger<FirmwareUpdateService>>()
-            ?? NullLogger<FirmwareUpdateService>.Instance;
-
-        // Bridge activation action: opened at the "Power cycle WINC" prompt to trigger the
-        // device's bridge-mode state machine right before the flash tool starts programming.
-        // By deferring APPLY (SYSTem:COMMunicate:LAN:APPLY) until this point we give the
-        // firmware a guaranteed promptResponseDelay window (2 s) to complete the WiFi
-        // deinit/reinit cycle and call wifi_serial_bridge_Init() before the flash tool
-        // issues its first serial bridge query.
-        Action bridgeActivationAction = () =>
-        {
-            _appLogger.Information($"Opening {portName} to send bridge activation commands.");
-            try
-            {
-                // USB CDC virtual ports ignore the baud rate; match the Core transport defaults.
-                using var port = new SerialPort(portName, 9600, Parity.None, 8, StopBits.One)
-                {
-                    DtrEnable = true,
-                    RtsEnable = false,
-                    WriteTimeout = 2000
-                };
-                port.Open();
-                // Brief pause to allow the DTR signal to be recognised by the firmware
-                // before we send commands (firmware checks isCdcHostConnected via DTR).
-                Thread.Sleep(200);
-                // Re-assert the FW-update-requested flag (idempotent; belt-and-suspenders).
-                port.Write("SYSTem:COMMUnicate:LAN:FWUpdate\n");
-                Thread.Sleep(100);
-                // Trigger the WiFi manager REINIT → bridge-mode state machine.
-                port.Write("SYSTem:COMMunicate:LAN:APPLY\n");
-                // Give the firmware a moment to enqueue the APPLY before we close the port.
-                Thread.Sleep(300);
-                _appLogger.Information("Bridge activation commands sent successfully.");
-            }
-            catch (Exception ex)
-            {
-                _appLogger.Warning($"Bridge activation failed for {portName}: {ex.Message}");
-            }
-        };
-
-        // Start from the shared firmware config so the bootloader HID timeouts stay aligned with
-        // the transport (keeps read and write windows symmetric), then layer on WiFi-specific
-        // settings. The WiFi flow flashes via the external WINC tool rather than the HID loop, so
-        // the bootloader timeout is inert here, but starting from CreateOptions() keeps every
-        // construction site uniform (issue #575).
-        var wifiOptions = FirmwareUpdateServiceConfig.CreateOptions();
-        // winc_flash_tool.cmd requires an explicit release version folder.
-        // Keep legacy argument profile used by shipped WINC tool bundle.
-        wifiOptions.WifiFlashToolArgumentsTemplate = $"/p {{port}} /d WINC1500 /v {wifiVersion} /k /e /i aio /w";
-        wifiOptions.WifiPortOverride = portName;
-        // After sending FWUpdate (flag-only, no APPLY), disconnect quickly so the
-        // COM port is free for the bridge activation raw write at the "Power cycle
-        // WINC" prompt.  The FWUpdate flag persists in firmware RAM until APPLY fires.
-        wifiOptions.PostLanFirmwareModeDelay = TimeSpan.FromMilliseconds(100);
-        // Give Windows a little more time to re-enumerate the UART before reconnect attempts.
-        wifiOptions.PostWifiReconnectDelay = TimeSpan.FromSeconds(3);
-
-        return new FirmwareUpdateService(
-            FirmwareUpdateServiceConfig.CreateBootloaderHidTransport(),
-            _firmwareDownloadService,
-            new WifiPromptDelayProcessRunner(
-                new ProcessExternalProcessRunner(),
-                promptResponseDelay: TimeSpan.FromSeconds(2),
-                bridgeActivationAction: bridgeActivationAction),
-            firmwareLogger,
-            options: wifiOptions);
-    }
-
-    private static string NormalizeWifiFirmwareVersion(string rawVersion)
-    {
-        var normalized = (rawVersion ?? string.Empty).Trim();
-        if (normalized.StartsWith("v", StringComparison.OrdinalIgnoreCase))
-        {
-            normalized = normalized[1..];
-        }
-
-        if (string.IsNullOrWhiteSpace(normalized))
-        {
-            throw new InvalidOperationException("WiFi firmware version metadata is missing.");
-        }
-
-        // Keep command argument safe even if an unexpected tag format appears.
-        var invalidChars = new[] { ' ', '\t', '\r', '\n', '"', '\'', ';' };
-        if (normalized.IndexOfAny(invalidChars) >= 0)
-        {
-            throw new InvalidOperationException($"Invalid WiFi firmware version tag '{rawVersion}'.");
-        }
-
-        return normalized;
-    }
-
-    private void HandleFirmwareUpdateException(FirmwareUpdateException exception)
-    {
-        HasErrorOccured = true;
-
-        var summary = $"Firmware update failed during '{exception.Operation}' ({exception.FailedState}).";
-        _appLogger.Error(exception, summary);
-
-        if (!string.IsNullOrWhiteSpace(exception.RecoveryGuidance))
-        {
-            _appLogger.Warning($"Firmware recovery guidance: {exception.RecoveryGuidance}");
-        }
-
-        var dialogMessage = summary;
-        if (!string.IsNullOrWhiteSpace(exception.RecoveryGuidance))
-        {
-            dialogMessage += $"{Environment.NewLine}{Environment.NewLine}Suggested recovery: {exception.RecoveryGuidance}";
-        }
-
-        ShowFirmwareErrorDialog(dialogMessage);
-    }
-
-    private void ShowFirmwareErrorDialog(string message)
-    {
-        void ShowDialog()
-        {
-            var errorDialogViewModel = new ErrorDialogViewModel(message);
-            _dialogService.ShowDialog<ErrorDialog>(this, errorDialogViewModel);
-        }
-
-        var dispatcher = Application.Current?.Dispatcher;
-        if (dispatcher == null)
-        {
-            ShowDialog();
-            return;
-        }
-
-        dispatcher.Invoke(ShowDialog);
-    }
-
-    private static string GetFirmwareDownloadDirectory()
-    {
-        var firmwareDirectory = Path.Combine(
-            App.DaqifiDataDirectory,
-            "Firmware",
-            "PIC32");
-        Directory.CreateDirectory(firmwareDirectory);
-        return firmwareDirectory;
-    }
-
-    private static string GetWifiDownloadDirectory()
-    {
-        var wifiDirectory = Path.Combine(
-            App.DaqifiDataDirectory,
-            "Firmware",
-            "WiFi");
-        Directory.CreateDirectory(wifiDirectory);
-        return wifiDirectory;
-    }
-
-    private static IFirmwareDownloadService CreateDefaultFirmwareDownloadService()
-    {
-        return new GitHubFirmwareDownloadService(new HttpClient());
-    }
-
-    private static IFirmwareUpdateService CreateDefaultFirmwareUpdateService(
-        IFirmwareDownloadService firmwareDownloadService,
-        ILogger<FirmwareUpdateService> logger)
-    {
-        return new FirmwareUpdateService(
-            FirmwareUpdateServiceConfig.CreateBootloaderHidTransport(),
-            firmwareDownloadService,
-            new ProcessExternalProcessRunner(),
-            logger,
-            options: FirmwareUpdateServiceConfig.CreateOptions());
     }
 
     #endregion
@@ -1191,7 +723,7 @@ public partial class DaqifiViewModel : ObservableObject
             }
         }
         ConnectionManager.Instance.Disconnect(deviceToDisconnect);
-        RemoveNotification(deviceToDisconnect);
+        _firmwareCoordinator.RemoveFirmwareNotification(deviceToDisconnect);
 
         if (deviceToDisconnect.Equals(SelectedDevice))
         {
@@ -1736,54 +1268,18 @@ public partial class DaqifiViewModel : ObservableObject
 
     #region Firmware version checking methods
 
-    private string _latestFirmwareVersion;
-    public string LatestFirmwareVersionText => _latestFirmwareVersion;
-    [RelayCommand]
-    public async Task GetFirmwareupdatationList()
-    {
-        var connectedDevices = ConnectionManager.Instance.ConnectedDevices;
-        if (connectedDevices.Count == 0)
-        {
-            return;
-        }
+    /// <summary>
+    /// The latest published firmware version, surfaced for display. The firmware coordinator
+    /// owns the value and performs the underlying version check via
+    /// <see cref="FirmwareUpdateCoordinator.RefreshFirmwareUpdatesAsync"/>.
+    /// </summary>
+    public string LatestFirmwareVersionText => _firmwareCoordinator.LatestFirmwareVersion;
 
-        try
-        {
-            var latestRelease = await _firmwareDownloadService.GetLatestReleaseAsync(includePreRelease: true);
-            _latestFirmwareVersion = latestRelease?.Version.ToString() ?? string.Empty;
-            OnPropertyChanged(nameof(LatestFirmwareVersionText));
-
-            if (latestRelease == null)
-            {
-                return;
-            }
-
-            foreach (var device in connectedDevices)
-            {
-                var updateCheck = await _firmwareDownloadService.CheckForUpdateAsync(
-                    device.DeviceVersion ?? string.Empty,
-                    includePreRelease: true);
-                var isOutdated = updateCheck.UpdateAvailable;
-                device.IsFirmwareOutdated = isOutdated;
-
-                if (!string.IsNullOrWhiteSpace(device.DeviceSerialNo))
-                {
-                    if (isOutdated)
-                    {
-                        AddNotification(device, latestRelease.Version.ToString());
-                    }
-                    else
-                    {
-                        RemoveNotification(device);
-                    }
-                }
-            }
-        }
-        catch (Exception ex)
-        {
-            _appLogger.Warning($"Failed to check firmware updates: {ex.Message}");
-        }
-    }
+    /// <summary>
+    /// Prunes notifications whose owning device is no longer connected. General-purpose
+    /// (covers both app-version and firmware notifications); firmware-specific add/remove
+    /// now lives in the firmware coordinator.
+    /// </summary>
     private void RemoveNotification()
     {
         foreach (var notification in NotificationList.ToList())
@@ -1799,70 +1295,12 @@ public partial class DaqifiViewModel : ObservableObject
         NotificationCount = NotificationList.Count;
     }
 
-    private void AddNotification(IStreamingDevice device, string latestFirmware)
-    {
-        var message = $"Device With Serial {device.DeviceSerialNo} has Outdated Firmware. Please Update to Version {latestFirmware}.";
-
-        var existingNotification = NotificationList.FirstOrDefault(n => n.DeviceSerialNo != null
-                                                                        && n.IsFirmwareUpdate
-                                                                        && n.DeviceSerialNo == device.DeviceSerialNo);
-
-        if (existingNotification == null)
-        {
-            NotificationList.Add(new Notifications
-            {
-                DeviceSerialNo = device.DeviceSerialNo,
-                Message = message,
-                IsFirmwareUpdate = true
-            });
-        }
-
-        NotificationCount = NotificationList.Count;
-    }
-    private void RemoveNotification(IStreamingDevice deviceToRemove)
-    {
-        if (deviceToRemove?.DeviceSerialNo == null)
-        {
-            return;
-        }
-
-        var notificationsToRemove = NotificationList
-            .FirstOrDefault(x => x.DeviceSerialNo != null && x.DeviceSerialNo == deviceToRemove.DeviceSerialNo && x.IsFirmwareUpdate);
-
-        if (notificationsToRemove != null)
-        {
-            NotificationList.Remove(notificationsToRemove);
-            NotificationCount = NotificationList.Count;
-        }
-    }
-
     #endregion
 
     [RelayCommand]
     private void OpenNotifications()
     {
         IsNotificationsOpen = true;
-    }
-
-    private void ShowUploadSuccessMessage()
-    {
-        void ShowDialog()
-        {
-            var successDialogViewModel =
-                new SuccessDialogViewModel("Firmware update completed successfully.");
-            _dialogService.ShowDialog<SuccessDialog>(this, successDialogViewModel);
-        }
-
-        var dispatcher = Application.Current?.Dispatcher;
-        if (dispatcher == null)
-        {
-            ShowDialog();
-            CloseFlyouts();
-            return;
-        }
-
-        dispatcher.Invoke(ShowDialog);
-        CloseFlyouts();
     }
 
     #endregion
@@ -1888,9 +1326,10 @@ public partial class DaqifiViewModel : ObservableObject
         {
             var SerailDeviceProperty = connectedDevice.GetType().GetProperty("DeviceVersion");
             var DeviceVersion = SerailDeviceProperty.GetValue(connectedDevice)?.ToString();
-            if (!string.IsNullOrEmpty(_latestFirmwareVersion))
+            var latestFirmwareVersion = _firmwareCoordinator.LatestFirmwareVersion;
+            if (!string.IsNullOrEmpty(latestFirmwareVersion))
             {
-                connectedDevice.IsFirmwareOutdated = VersionHelper.Compare(DeviceVersion, _latestFirmwareVersion) < 0;
+                connectedDevice.IsFirmwareOutdated = VersionHelper.Compare(DeviceVersion, latestFirmwareVersion) < 0;
             }
 
             ConnectedDevices.Add(connectedDevice);
@@ -2066,7 +1505,7 @@ public partial class DaqifiViewModel : ObservableObject
                 break;
         }
         CanToggleLogging = ActiveChannels.Count > 0;
-        _ = GetFirmwareupdatationList();
+        _ = _firmwareCoordinator.RefreshFirmwareUpdatesAsync();
         RemoveNotification();
     }
     public async Task<MessageDialogResult> ShowMessage(string title, string message, MessageDialogStyle dialogStyle)
@@ -2118,6 +1557,37 @@ public partial class DaqifiViewModel : ObservableObject
         IsNotificationsOpen = false;
         IsFirmwareUpdatationFlyoutOpen = false;
     }
+
+    #region IFirmwareUpdateHost implementation
+
+    // The firmware coordinator reaches the view model's bound progress/status properties
+    // (SelectedDevice, FirmwareFilePath, IsFirmwareUploading, UploadFirmwareProgress, etc.)
+    // through their existing public [ObservableProperty] members, which already satisfy the
+    // IFirmwareUpdateHost setters implicitly. Only the members below need an explicit bridge
+    // because they map onto differently-named state or onto desktop singletons the coordinator
+    // intentionally never touches directly.
+
+    /// <summary>
+    /// Devices the firmware version check iterates. Sourced from <see cref="ConnectionManager"/>
+    /// (not the UI-facing <see cref="ConnectedDevices"/> collection) to exactly preserve the
+    /// pre-refactor behavior.
+    /// </summary>
+    IReadOnlyList<IStreamingDevice> IFirmwareUpdateHost.ConnectedDevices =>
+        ConnectionManager.Instance.ConnectedDevices;
+
+    /// <summary>The shared, bound notification collection the coordinator mutates.</summary>
+    ObservableCollection<Notifications> IFirmwareUpdateHost.Notifications => NotificationList;
+
+    /// <summary>Routes the in-progress-update flag to the connection manager.</summary>
+    IStreamingDevice? IFirmwareUpdateHost.DeviceBeingUpdated
+    {
+        set => ConnectionManager.Instance.DeviceBeingUpdated = value;
+    }
+
+    /// <summary>Re-syncs the notification badge after the coordinator adds/removes a notification.</summary>
+    void IFirmwareUpdateHost.RefreshNotificationCount() => NotificationCount = NotificationList.Count;
+
+    #endregion
 
     private bool CanExportAllLoggingSession()
     {
